feat: Refactor argument parser, add initial module
The refactor allows the argument parser to simply take in an option
map and output an argument map, which boils down to known arguments
and the parsed output values. What that value is comes down to the
option map; the "parse" property must be a function that returns
something, and takes in the provided value. For instance, the input/
output parsers will return a function that will get you a file handle
(output might need to change here), and the dialect parser will
simply return the value (since it should be an acceptable option at
this point. What the parser won't do now, is penalise you for
providing options that don't exist in the map. As long as the
known options are valid, then the rest will be passed through for
modules to process with their own option maps. If any of the
options in the provided map are invalid, then execution will stop.

Valid modules are tested, and must provide a default exported
function, which will be the equivalent of its "main" function. As
long as the module is valid, the default function will be executed
and all arguments passed into it to parse as necessary. Modules do
not need to accept options, but they can. What they do with the
program arguments is up to them.