feat: Parse Bike content to simple JSON
The design of this module might want some refactoring at some point.
Problem really is this isn't a general purpose tool, and the input
format is supposed to remain static. As long as that is the case,
we should be alright...

Some less pleasant assumptions have gone into this module, versus
the rest of the program. The Bike input converted to XML produces
arrays everywhere, even when it doesn't make sense, but this is
just the nature of a generic parser.